requirements.txt
python-dotenv==0.20.0
yagmail==0.15.280
PyYAML==6.0
python-crontab==2.6.0
pytz==2022.1